Fix aoc-call-with-slim
We no longer do crazy `buffer-swap-text' and `with-silent-modifications' hacks.

Instead, we store the contents of the original buffer in a temporary buffer by
manually copying it with `copy-to-buffer'. Then, we write out the slim buffer to
the same file and invoke the compilation functions. Finally, we restore the
original by writing out the temporary buffer to the file.

This fixes the undo history getting clobbered as well as the auto-slim
functionality firing.